fix: cast int to size_t to quieten compilers.
In line 1831, a cast is required because `malloc()` expects a `size_t` as the
first argument and an `int` was provided. The cast is perfectly safe here, as
the number of tests can not be signed. Perhaps the better choice would have
been to declare `acutest_list_size_` as an `unsigned` type. GCC-13.1 with
-Wconversion complains about this.